Rename: ticket 관련 파일, enum 파일 entity, repository 패키지에서 ticket 관련 파일들을 ticket 패키지로 이동 enum 패키지 생성 후 status, type 패키지로 구분
