+++ src/main/java/shootingstar/var/entity/Auction.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
+import shootingstar.var.enums.type.AuctionType;
 
 @Entity
 @Getter
