+++ src/main/java/shootingstar/var/entity/ScheduledTask.java
@@ -10,6 +10,7 @@ import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import shootingstar.var.enums.type.TaskType;
 
 @Entity
 @Getter
