+++ src/main/java/shootingstar/var/entity/ticket/Ticket.java
@@ -1,4 +1,4 @@
-package shootingstar.var.entity;
+package shootingstar.var.entity.ticket;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -9,13 +9,15 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
-import jakarta.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import shootingstar.var.entity.Auction;
+import shootingstar.var.entity.BaseTimeEntity;
+import shootingstar.var.entity.User;
 
 @Entity
 @Getter
