+++ src/main/java/shootingstar/var/enums/status/TicketReportStatus.java
@@ -1,4 +1,4 @@
-package shootingstar.var.entity;
+package shootingstar.var.enums.status;
 
 public enum TicketReportStatus {
     REFUSAL, STANDBY, APPROVE
