+++ src/main/java/shootingstar/var/enums/type/TaskType.java
@@ -1,4 +1,4 @@
-package shootingstar.var.entity;
+package shootingstar.var.enums.type;
 
 public enum TaskType {
     STANDBY, COMPLETE, CANCEL
