+++ src/main/java/shootingstar/var/repository/ScheduledTaskRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import shootingstar.var.entity.ScheduledTask;
-import shootingstar.var.entity.TaskType;
+import shootingstar.var.enums.type.TaskType;
 
 public interface ScheduledTaskRepository extends JpaRepository<ScheduledTask, Long> {
     List<ScheduledTask> findByTaskType(TaskType taskType);
