+++ src/main/java/shootingstar/var/repository/ticket/TicketReportRepository.java
@@ -1,9 +1,9 @@
-package shootingstar.var.repository;
+package shootingstar.var.repository.ticket;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import shootingstar.var.entity.TicketReport;
+import shootingstar.var.entity.ticket.TicketReport;
 
 public interface TicketReportRepository extends JpaRepository<TicketReport, Long> {
     @Query("select tr from TicketReport tr where tr.ticket.ticketId = :ticketId and tr.ticketReportNickname = :ticketReportNickname")
