+++ src/main/java/shootingstar/var/repository/ticket/TicketRepository.java
@@ -1,9 +1,8 @@
-package shootingstar.var.repository;
+package shootingstar.var.repository.ticket;
 
 import java.util.Optional;
-import javax.swing.text.html.Option;
 import org.springframework.data.jpa.repository.JpaRepository;
-import shootingstar.var.entity.Ticket;
+import shootingstar.var.entity.ticket.Ticket;
 
 public interface TicketRepository extends JpaRepository<Ticket, Long> {
     Optional<Ticket> findByTicketUUID(String ticketUUID);
